requirements.txt
boto3>=1.34.0
numpy>=1.26.0
pydantic>=2.0.0
strands-agents>=1.0.0
requests>=2.31.0
